refactor!: map instead of set in policy attachment for_eaches (#28)
* chore: terraform fmt

* refactor!: map instead of set in for_each

Changing policy attachments to use `map`s instead of `set`s in their
`for_each` definition. This way the keys of the `for_each` won't cause
anything dynamic and are not prone to issues that we've seen arising
when this module is looped over in another module (see https://github.com/clowdhaus/terraform-for-each-unknown).

BREAKING CHANGE: This change will result in the replacement of the
affected policy attachments. This can be prevented by moving the
resources in the state manually or by `moved` blocks. We can't add
`moved` blocks to this module as the keys where the resources are moved
_from_ are dynamic (they reference the `iam_policy_prefix`).

* fix: remove references to "examples" from gitignore